I refactored the databasemanager class

+++ src/main/java/nwes/passwordmanager/DatabaseManager.java
@@ -46,14 +46,17 @@ public class DatabaseManager {
                 + "expiry_date TEXT NOT NULL, "
                 + "cvv TEXT NOT NULL, "
                 + "owner_name TEXT NOT NULL, "
+                + "card_pin TEXT NOT NULL, "
+                + "network_type TEXT NOT NULL, "
+                + "card_type TEXT NOT NULL, "
                 + "date_added TIMESTAMP DEFAULT CURRENT_TIMESTAMP);");
 
         stmt.execute("CREATE TABLE IF NOT EXISTS Wallets ("
                 + "id INTEGER PRIMARY KEY AUTOINCREMENT, "
                 + "resource TEXT NOT NULL, "
                 + "twelve_words TEXT NOT NULL, "
-                + "password TEXT NOT NULL, "
                 + "address TEXT NOT NULL, "
+                + "password TEXT NOT NULL, "
                 + "date_added TIMESTAMP DEFAULT CURRENT_TIMESTAMP);");
 
         stmt.execute("CREATE TABLE IF NOT EXISTS Links ("
@@ -85,25 +88,8 @@ public class DatabaseManager {
             System.out.println("❌ Database Write Error: " + e.getMessage());
         }
     }
-
-    public void writeLinkTodb(String link, String url) {
-        String sql = "INSERT INTO Links (resource, link) VALUES (?,?)";
-
-        try (Connection conn = DriverManager.getConnection(DB_URL);
-             PreparedStatement pstmt = conn.prepareStatement(sql)) {
-
-            pstmt.setString(1, link);
-            pstmt.setString(2, url);
-            pstmt.executeUpdate();
-            System.out.println("✅ Link successfully inserted: " + link + " and " + url);
-
-        } catch (SQLException e) {
-            System.out.println("❌ Database Write Error: " + e.getMessage());
-        }
-    }
-
-    public void writeCardTodb(String resource, String cardNumber, String expiryDate, String cvv, String ownerName, String systemType,LocalDateTime date){
-        String sql = "INSERT INTO Cards (resource, card_number, expiry_date, cvv, owner_name, date_added) VALUES (?, ?, ?, ?, ?, ?)";
+    public void writeCardTodb(String resource, String cardNumber, String expiryDate, String cvv, String ownerName, String cardPin, String networkType, String cardType, LocalDateTime date){
+        String sql = "INSERT INTO Cards (resource, card_number, expiry_date, cvv, owner_name, card_pin, network_type, card_type, date_added) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?)";
 
         try (Connection conn = DriverManager.getConnection(DB_URL);
              PreparedStatement pstmt = conn.prepareStatement(sql)) {
@@ -113,7 +99,10 @@ public class DatabaseManager {
             pstmt.setString(3, expiryDate);
             pstmt.setString(4, cvv);
             pstmt.setString(5, ownerName);
-            pstmt.setTimestamp(6, Timestamp.valueOf(date));
+            pstmt.setString(6, cardPin);
+            pstmt.setString(7, networkType);
+            pstmt.setString(8, cardType);
+            pstmt.setTimestamp(9, Timestamp.valueOf(date));
 
             pstmt.executeUpdate();
             System.out.println("✅ Card successfully inserted: " + resource + " on " + date);
@@ -122,16 +111,32 @@ public class DatabaseManager {
             System.out.println("❌ Database Write Error: " + e.getMessage());
         }
     }
+    public void writeLinkTodb(String resource, String link) {
+        String sql = "INSERT INTO Links (resource, link) VALUES (?,?)";
 
-    public void writeWalletTodb(String resource, String words, String password, LocalDateTime date){
-        String sql = "INSERT INTO Wallets (resource, twelve_words, password, address, date_added) VALUES (?, ?, ?, ?, ?)";
+        try (Connection conn = DriverManager.getConnection(DB_URL);
+             PreparedStatement pstmt = conn.prepareStatement(sql)) {
+
+            pstmt.setString(1, resource);
+            pstmt.setString(2, link);
+            pstmt.executeUpdate();
+            System.out.println("✅ Link successfully inserted: " + resource + " and " + link);
+
+        } catch (SQLException e) {
+            System.out.println("❌ Database Write Error: " + e.getMessage());
+        }
+    }
+
+    public void writeWalletTodb(String resource, String words, String address, String password, LocalDateTime date){
+        String sql = "INSERT INTO Wallets (resource, twelve_words, address, password, date_added) VALUES (?, ?, ?, ?, ?)";
 
         try (Connection conn = DriverManager.getConnection(DB_URL);
              PreparedStatement pstmt = conn.prepareStatement(sql)) {
 
             pstmt.setString(1, resource);
             pstmt.setString(2, words);
-            pstmt.setString(3, password);
+            pstmt.setString(3, address);
+            pstmt.setString(4, password);
             pstmt.setTimestamp(4, Timestamp.valueOf(date));
 
             pstmt.executeUpdate();
@@ -165,6 +170,33 @@ public class DatabaseManager {
 
         return accounts;
     }
+
+    public List<Card> getAllCards(){
+        List<Card> cards = new ArrayList<>();
+        String sql = "SELECT resource, card_number, expiry_date, cvv, owner_name, card_pin, network_type, card_type, date_added FROM Cards";
+
+        try (Connection conn = DriverManager.getConnection(DB_URL);
+             PreparedStatement pstmt = conn.prepareStatement(sql);
+             ResultSet rs = pstmt.executeQuery()){
+
+            while(rs.next()){
+                String resource = rs.getString("resource");
+                String cardNumber = rs.getString("card_number");
+                String expiryDate = rs.getString("expiry_date");
+                String cvv = rs.getString("cvv");
+                String ownerName = rs.getString("owner_name");
+                String cardPin = rs.getString("card_pin");
+                String networkType = rs.getString("network_type");
+                String cardType = rs.getString("card_type");
+//                LocalDateTime dateAdded = LocalDateTime.parse(rs.getString("date_added").replace(" ", "T"));
+
+                cards.add(new Card(resource, cardNumber, expiryDate, cvv, ownerName, cardPin, networkType, cardType, LocalDateTime.now()));
+            }
+        } catch (Exception e) {
+            System.out.println("❌ Database Read Error (Cards): " + e.getMessage());
+        }
+        return cards;
+    }
     public List<Link> getAllLinks(){
         List<Link> links = new ArrayList<>();
         String sql = "SELECT resource, link FROM Links";
@@ -182,32 +214,9 @@ public class DatabaseManager {
         }
         return links;
     }
-    public List<Card> getAllCards(){
-        List<Card> cards = new ArrayList<>();
-        String sql = "SELECT resource, card_number, expiry_date, cvv, owner_name, date_added FROM Cards";
-
-        try (Connection conn = DriverManager.getConnection(DB_URL);
-             PreparedStatement pstmt = conn.prepareStatement(sql);
-             ResultSet rs = pstmt.executeQuery()){
-
-            while(rs.next()){
-                String resource = rs.getString("resource");
-                String cardNumber = rs.getString("card_number");
-                String expiryDate = rs.getString("expiry_date");
-                String cvv = rs.getString("cvv");
-                String ownerName = rs.getString("owner_name");
-//                LocalDateTime dateAdded = LocalDateTime.parse(rs.getString("date_added").replace(" ", "T"));
-
-                cards.add(new Card(resource, cardNumber, expiryDate, cvv, ownerName, LocalDateTime.now()));
-            }
-        } catch (Exception e) {
-            System.out.println("❌ Database Read Error (Cards): " + e.getMessage());
-        }
-        return cards;
-    }
     public List<Wallet> getAllWallets() {
         List<Wallet> wallets = new ArrayList<>();
-        String sql = "SELECT resource, twelve_words, password, date_added FROM Wallets";
+        String sql = "SELECT resource, twelve_words, address, password, date_added FROM Wallets";
 
         try (Connection conn = DriverManager.getConnection(DB_URL);
              PreparedStatement pstmt = conn.prepareStatement(sql);
@@ -216,12 +225,13 @@ public class DatabaseManager {
             while (rs.next()) {
                 String resource = rs.getString("resource");
                 String wordsString = rs.getString("twelve_words");
+                String address = rs.getString("address");
                 String password = rs.getString("password");
 //                LocalDateTime dateAdded = LocalDateTime.parse(rs.getString("date_added").replace(" ", "T"));
 
                 String[] wordsArray = wordsString.split(",");
 
-                wallets.add(new Wallet(resource, wordsArray, password, LocalDateTime.now()));
+                wallets.add(new Wallet(resource, wordsArray, address, password, LocalDateTime.now()));
             }
         } catch (Exception e) {
             System.out.println("❌ Database Read Error (Wallets): " + e.getMessage());
@@ -244,8 +254,8 @@ public class DatabaseManager {
             System.out.println("Failed to update account: " + e.getMessage());
         }
     }
-    public void updateCard(Card card, String oldResource, String oldCardNumber) {
-        String sql = "UPDATE Cards SET resource = ?, card_number = ?, expiry_date = ?, cvv = ?, owner_name = ? WHERE resource = ? AND card_number = ?";
+    public void updateCard(Card card, String oldResource, String oldCardNumber, String oldCardName, String oldCardDate) {
+        String sql = "UPDATE Cards SET resource = ?, card_number = ?, expiry_date = ?, cvv = ?, owner_name = ?, card_pin = ?, network_type = ?, card_type =? WHERE resource = ? AND card_number = ? AND owner_name = ? AND expiry_date = ?";
 
         try(Connection conn = DriverManager.getConnection(DB_URL);
             PreparedStatement pstmt = conn.prepareStatement(sql)) {
@@ -255,8 +265,13 @@ public class DatabaseManager {
             pstmt.setString(3, card.getExpiryDate());
             pstmt.setString(4, card.getCvv());
             pstmt.setString(5, card.getOnwerName());
-            pstmt.setString(6, oldResource);
-            pstmt.setString(7, oldCardNumber);
+            pstmt.setString(6, card.getCardPincode());
+            pstmt.setString(7, card.getCardNetworkType());
+            pstmt.setString(8, card.getCardType());
+            pstmt.setString(9, oldResource);
+            pstmt.setString(10, oldCardNumber);
+            pstmt.setString(11, oldCardName);
+            pstmt.setString(12, oldCardDate);
             pstmt.executeUpdate();
 
         } catch (SQLException e) {
@@ -279,8 +294,8 @@ public class DatabaseManager {
             System.out.println("Failed to update link: " + e.getMessage());
         }
     }
-    public void updateWallet(Wallet wallet, String oldResource) {
-        String sql = "UPDATE Wallets SET resource = ?, twelve_words = ?, password = ? WHERE resource = ?";
+    public void updateWallet(Wallet wallet, String oldResource, String oldAddress) {
+        String sql = "UPDATE Wallets SET resource = ?, twelve_words = ?, address = ?, password = ? WHERE resource = ? AND address = ?";
         String wordText = String.join(",", wallet.getTwelveWords());
 
         try(Connection conn = DriverManager.getConnection(DB_URL);
@@ -288,8 +303,10 @@ public class DatabaseManager {
 
             pstmt.setString(1, wallet.getResource());
             pstmt.setString(2, wordText);
-            pstmt.setString(3, wallet.getPassword());
-            pstmt.setString(4, oldResource);
+            pstmt.setString(3, wallet.getAddress());
+            pstmt.setString(4, wallet.getPassword());
+            pstmt.setString(5, oldResource);
+            pstmt.setString(6, oldAddress);
             pstmt.executeUpdate();
 
         } catch (SQLException e) {
@@ -311,13 +328,15 @@ public class DatabaseManager {
         }
     }
     public void deleteCard(Card card) {
-        String sql = "DELETE FROM Cards WHERE resource = ? AND card_number = ?";
+        String sql = "DELETE FROM Cards WHERE resource = ? AND card_number = ? AND owner_name = ? AND expiry_date = ?";
 
         try(Connection conn = DriverManager.getConnection(DB_URL);
             PreparedStatement pstmt = conn.prepareStatement(sql)) {
 
             pstmt.setString(1, card.getResource());
             pstmt.setString(2, card.getCardNumber());
+            pstmt.setString(3, card.getOnwerName());
+            pstmt.setString(4, card.getExpiryDate());
             pstmt.executeUpdate();
 
         } catch (SQLException e) {
@@ -339,12 +358,13 @@ public class DatabaseManager {
         }
     }
     public void deleteWallet(Wallet wallet) {
-        String sql = "DELETE FROM Wallets WHERE resource = ?";
+        String sql = "DELETE FROM Wallets WHERE resource = ? AND address = ?";
 
         try(Connection conn = DriverManager.getConnection(DB_URL);
             PreparedStatement pstmt = conn.prepareStatement(sql)) {
 
             pstmt.setString(1, wallet.getResource());
+            pstmt.setString(2, wallet.getAddress());
             pstmt.executeUpdate();
 
         } catch (SQLException e) {
